+++ pages/sponsor/information.tsx
@@ -2,6 +2,8 @@ import SeoHeader from '@/components/layout/SeoHeader';
 import { Main } from '@/components/sponsor/information/Main';
 import { AboutPycon } from '@/components/sponsor/information/AboutPycon';
 import { Guide } from '@/components/sponsor/information/Guide';
+import SponsorLevelTable from '@/components/sponsor/information/SponsorLevelTable';
+import SponsorFAQToggle from '@/components/sponsor/information/SponsorFAQToggle';
 
 const SponsorInformation = () => {
   return (
@@ -10,6 +12,8 @@ const SponsorInformation = () => {
       <Main />
       <AboutPycon />
       <Guide />
+      <SponsorLevelTable />
+      <SponsorFAQToggle />
     </>
   );
 };
